+++ scripts/Groovy/kafka/KafkaPublisher.groovy
@@ -1,7 +1,7 @@
 @GrabResolver(root = 'https://packages.confluent.io/maven/', name = 'Confluent')
 @GrabConfig(systemClassLoader = true)
-@Grab(group = 'org.apache.kafka', module = 'kafka-clients', version = '2.3.1')
-@Grab(group = 'io.confluent', module = 'kafka-avro-serializer', version = '5.2.2')
+@Grab(group = 'org.apache.kafka', module = 'kafka-clients', version = '2.8.2')
+@Grab(group = 'io.confluent', module = 'kafka-avro-serializer', version = '5.5.15')
 @Grab(group = 'ch.qos.logback', module = 'logback-classic', version = '1.4.7')
 @Grab(group = 'com.jayway.jsonpath', module = 'json-path', version = '2.8.0')
 @Grab(group = 'io.micrometer', module = 'micrometer-registry-jmx', version = '1.5.7')
@@ -117,6 +117,7 @@ if (options.multiple) {
     count++
 }
 
+clearRequests()
 producer.close()
 LOGGER.info("Produced ${count} records")
 
@@ -165,7 +166,14 @@ void sendMessage(def producerRecord) {
     numberOfMessages.increment()
 
     if (requests.size() == 500) {
-        requests.each { it.get() }
-        requests.clear()
+        clearRequests()
     }
+}
+
+void clearRequests() {
+    requests.each { 
+        def metadata = it.get()
+        LOGGER.fine({ "Produced message to ${metadata.topic()} at ${metadata.partition()}:${metadata.offset()}".toString() }) 
+    }
+    requests.clear()
 }